Upgrade the user's exeperience (Read bellow)
I decide to re-implment the ``disabled`` function for the button because when a button is disabled, any error message appears (like the validation's error).

So I mix up the validation input (that's automatically created when a input is set like ``input="email"`` or the ``require`` tag is put) with a js validation that's add the new ``muted`` style.